change order of setting statusitem properties to fix cutoff
This still isn't entirely "fixed"; in some instances OS X will try to
fit it in a previously used but now blank space for some reason,
which more than likely won't fit the entire TypeStatus item. It
really seems like this is a bug that's out of the control of
TypeStatus and needs to be fixed by Apple. That, or I delve into
whatever process manages the menu bar, which I don't really feel like
doing...